+++ src/components/About.tsx
@@ -1,27 +1,32 @@
 import Image from 'next/image';
+import React from 'react';
 import styles from './About.module.css';
 
-const About = () => {
+const WorldProfile: React.FC = () => {
   return (
-    <section className={styles.about} id="about">
-      <div className={styles.inner}>
-        <h2 className={styles.title}>about</h2>
-        <div className={styles.profile}>
-          <p className={styles.profileImg}>
-            <Image src="/img/profile.jpg" alt="プロフィール画像" width={200} height={200} />
+    <div className={styles.container}>
+      <div className={styles.worldSection}>
+        <h1 className={styles.title}>Sho Katsumata`s RPG Portfolio</h1>
+      </div>
+      <div className={styles.profileSection}>
+        <div className={styles.profileImage}>
+          <Image src="/img/profile.jpg" alt="Sho Katsumata" width={200} height={200} />
+        </div>
+        <div className={styles.profileInfo}>
+          <h2 className={styles.profileName}>Sho Katsumata</h2>
+          <p className={styles.profileTitle}>Full-stack Developer</p>
+          <p className={styles.profileDescription}>
+            A passionate developer with a quest to create amazing web applications. Skilled in
+            various programming languages and always eager to learn more. Ready to take on any
+            coding challenge that comes my way!
           </p>
-          <div className={styles.profileBody}>
-            <p>
-              テキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入ります。
-            </p>
-            <p>
-              テキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入りますテキストが入ります。
-            </p>
-          </div>
         </div>
       </div>
-    </section>
+      <div className={styles.pixelCharacter}>
+        <Image src="/img/pixel-character.png" alt="Pixel Character" width={64} height={64} />
+      </div>
+    </div>
   );
 };
 
-export default About;
+export default WorldProfile;
