--- src/components/CharacterStats.module.css
@@ -1,69 +1,0 @@
-.container {
-  width: 100%;
-  max-width: 48rem;
-  padding: 1.5rem;
-  margin: 0 auto;
-  background-color: #000;
-  border: 2px solid #4caf50;
-  border-radius: 0.5rem;
-}
-
-.title {
-  margin-bottom: 1.5rem;
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: #4caf50;
-  text-align: center;
-}
-
-.chartContainer {
-  margin-bottom: 2rem;
-}
-
-.categoryContainer {
-  margin-bottom: 1.5rem;
-}
-
-.categoryTitle {
-  margin-bottom: 0.5rem;
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #4caf50;
-}
-
-.skillsGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 1rem;
-}
-
-.skillItem {
-  display: flex;
-  flex-direction: column;
-}
-
-.skillName {
-  margin-bottom: 0.25rem;
-  font-size: 0.875rem;
-  color: #fff;
-}
-
-.skillBar {
-  width: 100%;
-  height: 0.625rem;
-  overflow: hidden;
-  background-color: #4a4a4a;
-  border-radius: 9999px;
-}
-
-.skillBarFill {
-  height: 100%;
-  background-color: #4caf50;
-  border-radius: 9999px;
-}
-
-@media (width <= 640px) {
-  .skillsGrid {
-    grid-template-columns: 1fr;
-  }
-}
