+++ src/components/Contact.module.css
@@ -1,53 +1,43 @@
-.contact {
-  padding: 80px 0;
-  background-color: #fff;
+.messaging {
+  padding: 20px;
+  background-color: #111;
+  border: 2px solid #4caf50;
+  border-radius: 10px;
 }
 
-.inner {
-  max-width: 1000px;
-  padding: 0 20px;
-  margin: 0 auto;
-}
-
-.title {
-  margin-bottom: 40px;
-  font-size: 32px;
-  font-weight: bold;
+.messaging h2 {
+  margin-bottom: 20px;
+  color: #4caf50;
   text-align: center;
 }
 
-.lead {
-  margin-bottom: 40px;
-  text-align: center;
-}
-
-.contactList {
+.messageList {
   display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
+  flex-direction: column;
+  gap: 15px;
 }
 
-.contactItem {
-  margin: 0 20px;
-  font-size: 18px;
-  color: #333;
+.messageItem {
+  display: flex;
+  align-items: center;
+  padding: 15px;
+  color: #fff;
   text-decoration: none;
+  background-color: #222;
+  border: 1px solid #4caf50;
+  border-radius: 5px;
+  transition: transform 0.3s ease;
 }
 
-.contactItem:hover {
-  text-decoration: underline;
+.messageItem:hover {
+  transform: translateX(5px);
 }
 
-.spOnly {
-  display: none;
+.messageIcon {
+  margin-right: 15px;
+  font-size: 24px;
 }
 
-@media (width <= 767px) {
-  .spOnly {
-    display: inline;
-  }
-
-  .contactItem {
-    margin: 10px 20px;
-  }
+.messagePlatform {
+  font-size: 18px;
 }
