--- src/components/GameMenu.tsx
@@ -1,19 +1,0 @@
-import styles from './GameMenu.module.css';
-
-type GameMenuProps = {
-  setCurrentView: React.Dispatch<React.SetStateAction<string>>;
-};
-
-const GameMenu: React.FC<GameMenuProps> = ({ setCurrentView }) => {
-  return (
-    <nav className={styles.gameMenu}>
-      <button onClick={() => setCurrentView('world')}>World</button>
-      <button onClick={() => setCurrentView('stats')}>Skills</button>
-      <button onClick={() => setCurrentView('timeline')}>Timeline</button>
-      <button onClick={() => setCurrentView('quests')}>Works</button>
-      <button onClick={() => setCurrentView('messages')}>Contact</button>
-    </nav>
-  );
-};
-
-export default GameMenu;
