+++ src/components/Header.module.css
@@ -1,23 +1,25 @@
-.header {
-  padding: 20px 0;
-  background-color: #fff;
-  box-shadow: 0 2px 4px rgb(0 0 0 / 10%);
+.gameMenu {
+  display: flex;
+  justify-content: space-around;
+  padding: 10px;
+  background-color: #111;
+  border-top: 2px solid #4caf50;
 }
 
-.headerInner {
-  max-width: 1000px;
-  padding: 0 20px;
-  margin: 0 auto;
+.gameMenu button {
+  display: inline-block;
+  padding: 10px 20px;
+  margin: 4px 2px;
+  font-family: 'Press Start 2P', cursive;
+  font-size: 16px;
+  color: white;
   text-align: center;
+  text-decoration: none;
+  cursor: pointer;
+  background-color: #4caf50;
+  border: none;
 }
 
-.headerTitle {
-  margin-bottom: 5px;
-  font-size: 36px;
-  font-weight: bold;
-}
-
-.headerSubtitle {
-  font-size: 18px;
-  color: #666;
+.gameMenu button:hover {
+  background-color: #45a049;
 }
