+++ src/components/Header.tsx
@@ -1,14 +1,19 @@
 import styles from './Header.module.css';
 
-const Header = () => {
+type GameMenuProps = {
+  setCurrentView: React.Dispatch<React.SetStateAction<string>>;
+};
+
+const GameMenu: React.FC<GameMenuProps> = ({ setCurrentView }) => {
   return (
-    <header className={styles.header}>
-      <div className={styles.headerInner}>
-        <p className={styles.headerTitle}>Sho Katsumata</p>
-        <p className={styles.headerSubtitle}>Portfolio</p>
-      </div>
-    </header>
+    <nav className={styles.gameMenu}>
+      <button onClick={() => setCurrentView('world')}>About</button>
+      <button onClick={() => setCurrentView('stats')}>Skills</button>
+      <button onClick={() => setCurrentView('timeline')}>Timeline</button>
+      <button onClick={() => setCurrentView('quests')}>Works</button>
+      <button onClick={() => setCurrentView('messages')}>Contact</button>
+    </nav>
   );
 };
 
-export default Header;
+export default GameMenu;
