--- src/components/Messaging.module.css
@@ -1,43 +1,0 @@
-.messaging {
-  padding: 20px;
-  background-color: #111;
-  border: 2px solid #4caf50;
-  border-radius: 10px;
-}
-
-.messaging h2 {
-  margin-bottom: 20px;
-  color: #4caf50;
-  text-align: center;
-}
-
-.messageList {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.messageItem {
-  display: flex;
-  align-items: center;
-  padding: 15px;
-  color: #fff;
-  text-decoration: none;
-  background-color: #222;
-  border: 1px solid #4caf50;
-  border-radius: 5px;
-  transition: transform 0.3s ease;
-}
-
-.messageItem:hover {
-  transform: translateX(5px);
-}
-
-.messageIcon {
-  margin-right: 15px;
-  font-size: 24px;
-}
-
-.messagePlatform {
-  font-size: 18px;
-}
