--- src/components/Navigation.module.css
@@ -1,44 +1,0 @@
-.gnav {
-  padding: 10px 0;
-  background-color: #333;
-}
-
-.gnavInner {
-  max-width: 1000px;
-  padding: 0 20px;
-  margin: 0 auto;
-}
-
-.gnavList {
-  display: flex;
-  justify-content: center;
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-.gnavItem {
-  margin: 0 15px;
-}
-
-.gnavItem a {
-  font-size: 16px;
-  color: #fff;
-  text-decoration: none;
-  text-transform: uppercase;
-}
-
-.gnavItem a:hover {
-  text-decoration: underline;
-}
-
-@media (width <= 767px) {
-  .gnavList {
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .gnavItem {
-    margin: 10px 0;
-  }
-}
