--- src/components/QuestBoard.module.css
@@ -1,42 +1,0 @@
-.questBoard {
-  padding: 20px;
-  background-color: #111;
-  border: 2px solid #4caf50;
-  border-radius: 10px;
-}
-
-.questBoard h2 {
-  margin-bottom: 20px;
-  color: #4caf50;
-  text-align: center;
-}
-
-.questList {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 20px;
-}
-
-.questItem {
-  padding: 15px;
-  color: #fff;
-  text-decoration: none;
-  background-color: #222;
-  border: 1px solid #4caf50;
-  border-radius: 5px;
-  transition: transform 0.3s ease;
-}
-
-.questItem:hover {
-  transform: translateY(-5px);
-}
-
-.questItem h3 {
-  margin-bottom: 10px;
-  color: #4caf50;
-}
-
-.questItem p {
-  margin: 5px 0;
-  font-size: 14px;
-}
