+++ src/components/Skills.module.css
@@ -1,55 +1,69 @@
-.skills {
-  padding: 4rem 2rem;
-  background-color: #f8f9fa;
+.container {
+  width: 100%;
+  max-width: 48rem;
+  padding: 1.5rem;
+  margin: 0 auto;
+  background-color: #000;
+  border: 2px solid #4caf50;
+  border-radius: 0.5rem;
 }
 
-.skills h2 {
-  margin-bottom: 2rem;
-  font-size: 2.5rem;
-  color: #333;
+.title {
+  margin-bottom: 1.5rem;
+  font-size: 1.5rem;
+  font-weight: bold;
+  color: #4caf50;
   text-align: center;
 }
 
-.skillsList {
+.chartContainer {
+  margin-bottom: 2rem;
+}
+
+.categoryContainer {
+  margin-bottom: 1.5rem;
+}
+
+.categoryTitle {
+  margin-bottom: 0.5rem;
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: #4caf50;
+}
+
+.skillsGrid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 1rem;
+}
+
+.skillItem {
   display: flex;
-  flex-wrap: wrap;
-  gap: 1.5rem;
-  justify-content: center;
-  padding: 0;
-  list-style-type: none;
+  flex-direction: column;
 }
 
-.skillsList li {
-  padding: 1rem 1.5rem;
-  font-size: 1rem;
-  color: #555;
-  background-color: #fff;
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgb(0 0 0 / 10%);
-  transition: all 0.3s ease;
+.skillName {
+  margin-bottom: 0.25rem;
+  font-size: 0.875rem;
+  color: #fff;
 }
 
-.skillsList li:hover {
-  box-shadow: 0 4px 8px rgb(0 0 0 / 15%);
-  transform: translateY(-5px);
+.skillBar {
+  width: 100%;
+  height: 0.625rem;
+  overflow: hidden;
+  background-color: #4a4a4a;
+  border-radius: 9999px;
 }
 
-@media (width <= 768px) {
-  .skills {
-    padding: 3rem 1rem;
-  }
+.skillBarFill {
+  height: 100%;
+  background-color: #4caf50;
+  border-radius: 9999px;
+}
 
-  .skills h2 {
-    font-size: 2rem;
-  }
-
-  .skillsList {
-    gap: 1rem;
-  }
-
-  .skillsList li {
-    padding: 0.75rem 1rem;
-    font-size: 0.9rem;
+@media (width <= 640px) {
+  .skillsGrid {
+    grid-template-columns: 1fr;
   }
 }
